requirements.txt
bokeh
matplotlib
numpy
openpyxl
pandas
pyomo
scipy
spyder
xlrd
ramanspy
python-dotenv